Organize project structure with dedicated folders for models, services, exceptions and update readme.
