+++ Exceptions/InventoryException.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace InventorySystem.Exceptions
+{
+    public class InventoryException : Exception
+    {
+        public InventoryException(string message) : base(message) { }
+    }
+}
