+++ Exceptions/ProductAlreadyExistsException.cs
@@ -1,0 +1,7 @@
+namespace InventorySystem.Exceptions
+{
+    public class ProductAlreadyExistsException : InventoryException
+    {
+        public ProductAlreadyExistsException(string message) : base(message) { }
+    }
+}
