+++ README.md
@@ -1,2 +1,25 @@
 # Inventory-Control-System
-Here is a sketch of a small inventory control system made with C# .net. The system is embryonic and there will be improvements.
+Sistema de Controle de Inventário
+
+Este projeto é um sistema de controle de inventário simples desenvolvido em C# .NET. Ele permite o gerenciamento de produtos e o registro de movimentos de inventário.
+
+## Funcionalidades
+
+- Adicionar e atualizar produtos
+- Registrar entradas e saídas de inventário
+- Validar existência de produtos, categorias e fornecedores
+
+## Tecnologias Utilizadas
+
+- C# .NET
+- Entity Framework Core para interações com o banco de dados
+- ASP.NET Core (opcional) para uma possível interface de API
+
+## Instalação
+
+Para executar este projeto, você precisará do .NET SDK instalado em sua máquina. Siga estas etapas:
+
+1. Clone o repositório para sua máquina local:
+
+```bash
+git clone <URL-do-repositório>
