+++ Services/IInventoryService.cs
@@ -1,0 +1,10 @@
+using InventorySystem.Models;
+using System.Threading.Tasks;
+
+namespace InventorySystem.Services
+{
+    public interface IInventoryService
+    {
+        Task RecordInventoryMovementAsync(InventoryMovement movement);
+    }
+}
