+++ Services/ProductService.cs
@@ -1,0 +1,34 @@
+using InventorySystem.Models;
+using InventorySystem.DbContext;
+using System.Threading.Tasks;
+using System; // Para ArgumentNullException
+using System.Linq.Expressions; // Para Expression<Func<T, bool>>
+
+namespace InventorySystem.Services
+{
+    public class ProductService : IProductService
+    {
+        private readonly IDbContext _dbContext;
+
+        public ProductService(IDbContext dbContext)
+        {
+            _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
+        }
+
+        public async Task AddProductAsync(Product newProduct)
+        {
+            // Falta implementar
+        }
+
+        public async Task UpdateProductAsync(Product updatedProduct)
+        {
+            // Falta implementar
+        }
+
+        private async Task<bool> EntityExistsAsync<T>(Expression<Func<T, bool>> predicate) where T : class
+        {
+            // Implementação depende do ORM específico, como EF Core
+            return await _dbContext.Set<T>().AnyAsync(predicate);
+        }
+    }
+}
