--- inventory control system/Inventory-Control-System/README.md
@@ -1,2 +1,0 @@
-# Inventory-Control-System
-Here is a sketch of a small inventory control system made with C# .net. The system is embryonic and there will be improvements.
